requirements.txt
torch
torchaudio
pyannote.audio
pyyaml
tqdm
numpy
pandas